Preliminary support for wiki
Some new stuff for wikis as requested.  I reverted the exposing of the
WebAgent, although it might be a good idea to leave it open since it
adds flexibility at no overhead.

I think everything is here, except wiki page revision diffs which i
don't think have support in the public api. Everything is in a new Wiki
object in the Subreddit class.

The only modification to existing code is in the Thing class.  I had to
add parsing by generic type (Parse<T>) so I could get Listing support
for Things that aren't "Things" like page revisions.  Slight change to
Listing to call the generic Parse function. If this is bad form let me
know.  It shouldn't affect anything.

I also have an xunit test project that I didn't check in for this.  Let
me know if you want that too.