Listing - adding maximum listing results for retrieve
Updating the ListingEnumerator class to handle a maximum limit for
retrieving listings.

Updating the Listing class with comments and adding overloaded
GetEnumerator methods to specify the maximum number of listings and
number of listings per request to return.

Adding GetListing methods to provide an easy way to get convert the
ListingEnumerator to an IEnumerable with the maximum and per response
listing parameters