Allow precompiled binaries in "Release" folder
Please keep the "no binary" crap to the linux world. In Windows we provide binaries (You build them anyways).
Changing this line allows binaries from the "Release" folder to be added to the repository. If you want people to use your library, actually provide it.

I could not use the library because I lack the required visual studio version.