Fixing parameters passed to set flair
Subreddit is encoded into the url supplied, not as another parameter in the json.

The name of the user making the change is also a required parameter. Using the
user field that reddit claims to be logged in as for this value.

This is based on the documentation at https://www.reddit.com/dev/api#POST_api_flair.